+++ test/pow/store/backend/mnesia_cache_test.exs
@@ -315,6 +315,23 @@ defmodule Pow.Store.Backend.MnesiaCacheTest do
       :rpc.call(node_a, MnesiaCache.Unsplit, :__heal__, [node_b, [flush_tables: :all]])
       assert :rpc.call(node_a, :mnesia, :system_info, [:running_db_nodes]) == [node_b, node_a]
     end
+
+    test "handles `extra_db_nodes: {module, function, arguments}`" do
+      :mnesia.kill()
+
+      # Init node a and write to it
+      node_a = spawn_node("a")
+      {:ok, _pid} = :rpc.call(node_a, Supervisor, :start_child, [Pow.Supervisor, {MnesiaCache, @default_config}])
+      assert :rpc.call(node_a, :mnesia, :system_info, [:extra_db_nodes]) == []
+      assert :rpc.call(node_a, :mnesia, :system_info, [:running_db_nodes]) == [node_a]
+
+      # Join cluster with node b
+      node_b = spawn_node("b")
+      config = @default_config ++ [extra_db_nodes: {Node, :list, []}]
+      {:ok, _pid} = :rpc.call(node_b, Supervisor, :start_child, [Pow.Supervisor, {MnesiaCache, config}])
+      assert :rpc.call(node_b, :mnesia, :system_info, [:extra_db_nodes]) == [node_a]
+      assert :rpc.call(node_b, :mnesia, :system_info, [:running_db_nodes]) == [node_a, node_b]
+    end
   end
 
   defp spawn_node(sname) do