+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## v1.0.24 (TBA)
 
-* [`Pow.Store.Backend.MnesiaCache`] Now accepts `extra_db_nodes: :all` and `extra_db_nodes: {module, function, arguments}` to automatically connect to all visible nodes
-
+* [`Pow.Store.Backend.MnesiaCache`] Now accepts `extra_db_nodes: {module, function, arguments}` to fetch nodes when MnesiaCache starts up
 
 ## v1.0.23 (2021-03-22)