+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.0.24 (TBA)
+
+* [`Pow.Store.Backend.MnesiaCache`] Now accepts `extra_db_nodes: {module, function, arguments}` to fetch nodes when MnesiaCache starts up
+
 ## v1.0.23 (2021-03-22)
 
 ### Enhancements